backend/requirements.txt
fastapi
uvicorn
pydantic
pymongo
python-dotenv
python-jose
google-auth
requests
bcrypt
openai
python-multipart